+++ app/utils/permissions.server.ts
@@ -1,5 +1,6 @@
 import { json } from "@remix-run/cloudflare";
 import { WorkerDB } from "lib/db.ts";
+import { UserRepository } from "repositories/user.ts";
 import { requireUserId } from "./auth.server.ts";
 import { getAuthSessionStorage } from "./session.server.ts";
 
@@ -18,24 +19,11 @@ export async function requireUserWithPermission(
 ) {
   const userId = await requireUserId(authSessionStorage, db, request);
   const permissionData = parsePermissionString(permission);
-  const user = await db
-    .selectFrom("person")
-    .selectAll()
-    .where("person.id", "=", userId)
-    .innerJoin("role_person", "role_person.person_id", "person.id")
-    .innerJoin("role", "role.id", "role_person.role_id")
-    .innerJoin("role_permission", "role_permission.role_id", "role.id")
-    .innerJoin("permission", "permission.id", "role_permission.permission_id")
-    .where((eb) =>
-      eb.and([
-        eb("permission.action", "=", permissionData.action),
-        eb("permission.entity", "=", permissionData.entity),
-        permissionData.access
-          ? eb("permission.access", "in", permissionData.access)
-          : eb.val(true),
-      ])
-    )
-    .executeTakeFirst();
+  const user = await UserRepository.getUserWithPermissions(
+    db,
+    userId,
+    permissionData
+  );
   if (!user) {
     throw json(
       {
@@ -49,7 +37,7 @@ export async function requireUserWithPermission(
   return user.id;
 }
 
-function parsePermissionString(permissionString: PermissionString) {
+export function parsePermissionString(permissionString: PermissionString) {
   const [action, entity, access] = permissionString.split(":") as [
     Action,
     Entity,
